move models to models folder
