+++ testipy/lib_modules/browser_manager/playwright.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python3
 import base64
 
 from time import sleep
