+++ testipy/lib_modules/py_inspector.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 import os
 import importlib.util
 import inspect
