+++ testipy/lib_modules/webhook_http_listener.py
@@ -1,5 +1,3 @@
-#!/usr/bin/env python3
-
 import sys
 import threading
 import json
