+++ testipy/models/__init__.py
@@ -1,0 +1,2 @@
+from testipy.models.attr import PackageAttr, SuiteAttr, TestMethodAttr
+from testipy.models.package_manager import PackageManager, PackageDetails, SuiteDetails, TestDetails
