+++ testipy/reporter/__init__.py
@@ -1,3 +1,2 @@
 from testipy.reporter.report_interfaces import ReportInterface
 from testipy.reporter.report_manager import ReportManager
-from testipy.reporter.package_manager import PackageManager, PackageDetails, SuiteDetails, TestDetails
